# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/trips": {
      "get": {
        "operationId": "PlanningController_listTrips",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Viajes del alcance (paginación SQL, facetas por estado, prioridad, servicio y condición).",
        "tags": [
          "planning"
        ]
      },
      "post": {
        "operationId": "PlanningController_createTrip",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Planifica un viaje (RN-002 sin superposición; recursos elegibles; servicio vigente opcional).",
        "tags": [
          "planning"
        ]
      }
    },
...
    "/api/v1/driver/actions": {
      "post": {
        "operationId": "DriverController_submit",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Acción del outbox offline. Respuesta Confirmada | Rechazada (definitiva, con motivo).",
        "tags": [
          "driver"
        ]
      }
    },
    "/api/v1/cargo-shipments": {
      "get": {
        "operationId": "CargoController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cargo"
        ]
      },
      "post": {
        "operationId": "CargoController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CargoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "cargo"
        ]
      }
    },
...
    "/api/v1/cargo-shipments/{id}": {
      "patch": {
        "operationId": "CargoController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CargoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cargo"
        ]
      }
    },
    "/api/v1/cargo-shipments/{id}/assign": {
      "post": {
        "operationId": "CargoController_assign",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cargo"
        ]
      }
    },
    "/api/v1/cargo-shipments/{id}/advance": {
      "post": {
        "operationId": "CargoController_advance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cargo"
        ]
      }
    },
...
    "/api/v1/files": {
      "post": {
        "operationId": "FilesController_reserve",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReserveDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Reserva un archivo y devuelve una URL temporal de subida directa (10 min).",
        "tags": [
          "files"
        ]
      }
    },
    "/api/v1/files/{id}/complete": {
      "post": {
        "operationId": "FilesController_complete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "files"
        ]
      }
    },
    "/api/v1/files/{id}/download-url": {
      "get": {
        "operationId": "FilesController_downloadUrl",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "URL temporal de vista/descarga (5 min). El acceso queda auditado.",
        "tags": [
          "files"
        ]
      }
    },
    "/api/v1/files/blob/{token}": {
      "put": {
        "operationId": "FilesController_putBlob",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "files"
        ]
      },
      "get": {
        "operationId": "FilesController_getBlob",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "files"
        ]
      }
    },
...
  },
  "info": {
    "title": "Sistema de Gestión de Transportes — API",
    "description": "DOC-E-BE · Contratos API definitivos (REST /api/v1). Errores application/problem+json (ADR-013). Autorización por permiso + alcance + política; el tenant se resuelve SIEMPRE desde la sesión.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "TripDto": {
        "type": "object",
        "properties": {}
      },
...
      "AdvanceDto": {
        "type": "object",
        "properties": {}
      },
      "DriverActionDto": {
        "type": "object",
        "properties": {}
      },
      "CargoDto": {
        "type": "object",
        "properties": {}
      },
      "AssignDto": {
        "type": "object",
        "properties": {}
      },
...
      "ReserveDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}